Move some type definitions from store/controller to store/proto
