+++ frontend/src/modules/app/OverlayScreen.tsx
@@ -15,7 +15,7 @@ import {
   getHasActiveAlarms,
   getScreenStatusLock,
 } from '../../store/controller/selectors';
-import { MessageType } from '../../store/controller/types';
+import { MessageType } from '../../store/proto/types';
 
 import ModalPopup from '../controllers/ModalPopup';
 import MultiStepWizard from '../displays/MultiStepWizard';
