+++ frontend/src/modules/info/FiO2Info.tsx
@@ -9,7 +9,7 @@ import React from 'react';
 import { connect, useDispatch } from 'react-redux';
 import { createStructuredSelector } from 'reselect';
 import { ParametersRequest } from '../../store/proto/mcu_pb';
-import { MessageType } from '../../store/controller/types';
+import { MessageType } from '../../store/proto/types';
 import { commitRequest, commitDraftRequest } from '../../store/controller/actions';
 import { getParametersFiO2, getSensorMeasurementsFiO2 } from '../../store/controller/selectors';
 import { StoreState } from '../../store/types';
