+++ frontend/src/modules/settings/SettingsPage.tsx
@@ -10,7 +10,7 @@ import React, { useCallback } from 'react';
 import { useDispatch } from 'react-redux';
 import { commitRequest } from '../../store/controller/actions';
 import { FrontendDisplaySetting, SystemSettingRequest } from '../../store/proto/frontend_pb';
-import { MessageType } from '../../store/controller/types';
+import { MessageType } from '../../store/proto/types';
 import { a11yProps, TabPanel } from '../controllers/TabPanel';
 import { DisplayTab, InfoTab, TestCalibrationTab } from './containers';
