+++ frontend/src/store/controller/reducers/backend.ts
@@ -7,9 +7,8 @@ import {
   LogEventType,
 } from '../../proto/mcu_pb';
 import { RotaryEncoder } from '../../proto/frontend_pb';
+import { MessageType, PBMessage } from '../../proto/types';
 import {
-  MessageType,
-  PBMessage,
   StateUpdateAction,
   EventLog,
   RotaryEncoderParameter,
